backend/requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
pillow
python-multipart
